multiple stories based on user input
Created two functions which convert a string to an integer, and check if an integer is within a range. These are used when making the user pick a number from 1-100, which is used to determine the story that the user will see. I also added multiple loops which ensure the script only progresses when the user has entered something valid.